backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.13
python-multipart==0.0.9
langchain==0.2.16
langchain-openai==0.1.20
langchain-core==0.2.38
openai==1.37.1
httpx==0.27.2
pypdf==5.5.0
tqdm==4.67.1
langchain-community
faiss-cpu
langchain-openai